chore(refactor): extract sentry initialization into its own file
